+++ PIGGISWS/Models/DFactura.cs
@@ -5,7 +5,7 @@ namespace PIGGISWS.Models;
 public class DFactura
 {
     public int DFAC_EMPRESA {  get; set; }
-    public long DFAC_CFAC_COMPROBA {  get; set; }
+    public decimal DFAC_CFAC_COMPROBA {  get; set; }
     public decimal DFAC_CANTIDAD { get; set; }
     public decimal DFAC_TOTAL {  get; set; }
     public int DFAC_SECUENCIA { get; set; }
